+++ src/scripts/server/player-prefs/ui.ts
@@ -28,6 +28,7 @@ const COVERAGE = [
     "obsidian",
     "ores",
     "plants",
+    "rocksAbundant",
     "rocksCommon",
     "rocksUncommon",
     "sandLike",